rename github_utils
